requirements.txt
cmake==3.30.0
decorator==5.1.1
dlib==19.24.99
face-recognition==1.3.0
face-recognition-models==0.3.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
imageio-ffmpeg==0.5.1
numpy==2.0.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
pillow_heif==0.17.0
psycopg2-binary==2.9.9
requests==2.31.0
Wand==0.6.13
python-dotenv
